refactor: Node class에서 dummyNode 생성 방식 변경
* Node class를 LinkedList 파일로 이동
